requirements.txt
# Core dependencies
numpy>=1.24.0
cupy-cuda12x>=12.0.0  # For CUDA 12.x (RTX 3060)

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Data handling
pandas>=2.0.0

# Testing
pytest>=7.3.0
pytest-benchmark>=4.0.0

# Utilities
tqdm>=4.65.0  # Progress bars

# Optional: Jupyter for analysis
jupyter>=1.0.0
ipykernel>=6.23.0
